refactor: adjusted name
